[decision-ms] Implement RoleManager and Add Goalkeeper Role (#91)
#### RoleManager
Its main responsibility is to choose which roles should be executing at
any given time. This pull request adds the mechanisms to enable this
class to choose role classes based on private heuristics.

#### Goalkeeper role
As an example, a goalkeeper role was created. It is managed by the
`RoleManager` and produces `BehaviorCandidates`.

#### Changes
This pull requests also changes the structure of the
`decision::processor`, moving world update implementations inside
`World` class and takes `RoleManager` out of `Coach`, since the only
dependency between these two classes is the `World`, which can be freely
shared.